mixtral_cpu/requirements.txt
fastapi
pydantic~=2.7.0
uvicorn
loguru
huggingface-hub
llama-cpp-python